Fix the handling of _.md pages
The _.md pages use the url in the front matter to show up in place of
the index pages. That seems to only work reliable only when disabling
the 'section'. The 'taxonomy' kind is also disabled to avoid a Hugo
warning and the 'error-disable-taxonomy' is use to avoid another warning
about 'taxonomy'.